# apps/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "manhwa-backend",
    "description": "Ikiru Bot manhwa scraper API. Backend runs fully on local VPS Postgres. Use Bearer token for protected endpoints. RSS feeds support dual-read params: format (manhwa/manhua/manga) + type (legacy alias), country (KR/CN/JP) + origin (legacy alias).",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/catalog/search": {
      "get": {
        "summary": "Catalog Search",
        "operationId": "catalog_search_api_v1_catalog_search_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/catalog/stats": {
      "get": {
        "summary": "Catalog Stats",
        "operationId": "catalog_stats_api_v1_catalog_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/catalog": {
      "get": {
        "summary": "Catalog List",
        "operationId": "catalog_list_api_v1_catalog_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/catalog/{title_key}": {
      "get": {
        "summary": "Catalog Item",
        "operationId": "catalog_item_api_v1_catalog__title_key__get",
        "parameters": [
          {
            "name": "title_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Title Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/audit-log": {
      "get": {
        "summary": "Get Audit Log",
        "operationId": "get_audit_log_api_v1_audit_log_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/audit/stale": {
      "get": {
        "summary": "Audit Stale",
        "operationId": "audit_stale_api_v1_audit_stale_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "https://scanner.aldifhr.fun",
      "description": "Production (VPS)"
    }
  ]
}